+++ data_safe_haven/config/__init__.py
@@ -5,7 +5,8 @@ from .config import (
     ConfigSectionSRE,
     ConfigSubsectionRemoteDesktopOpts,
 )
-from .pulumi import DSHPulumiConfig, DSHPulumiProject
+from .pulumi import DSHPulumiConfig
+from .pulumi_project import DSHPulumiProject
 
 __all__ = [
     "Config",
